refactor: extract Connor's types into separated packages
This commit extracts types used by Connor into separate package to prevents us from cycle import issue.
The main reason is to be able to use `connor.Deal` instead of `sonm.Deal` in the antifraud package.
This is required because `connor.Deal` have an info about the main benchmark value which should be controlled by the anti-fraud.
