+++ connor/types/benchmarks.go
@@ -1,0 +1,52 @@
+package types
+
+import "github.com/sonm-io/core/proto"
+
+type Benchmarks sonm.Benchmarks
+
+func newZeroBenchmarks() Benchmarks {
+	return Benchmarks{Values: make([]uint64, sonm.MinNumBenchmarks)}
+}
+
+func (b *Benchmarks) setGPUEth(v uint64) {
+	b.Values[9] = v
+}
+
+func (b *Benchmarks) setGPUZcash(v uint64) {
+	b.Values[10] = v
+}
+
+func (b *Benchmarks) setGPURedshift(v uint64) {
+	b.Values[11] = v
+}
+
+func (b *Benchmarks) unwrap() *sonm.Benchmarks {
+	v := sonm.Benchmarks(*b)
+	return &v
+}
+
+func (b *Benchmarks) toMap() map[string]uint64 {
+	// warn: this is shitty crutch, but we should refactor
+	// CreateOrder method to omit this.
+
+	v := b.unwrap()
+	return map[string]uint64{
+		"cpu-sysbench-multi":  v.CPUSysbenchMulti(),
+		"cpu-sysbench-single": v.CPUSysbenchOne(),
+		"cpu-cores":           v.CPUCores(),
+		"ram-size":            v.RAMSize(),
+		"storage-size":        v.StorageSize(),
+		"net-download":        v.NetTrafficIn(),
+		"net-upload":          v.NetTrafficOut(),
+		"gpu-count":           v.GPUCount(),
+		"gpu-mem":             v.GPUMem(),
+		"gpu-eth-hashrate":    v.GPUEthHashrate(),
+		"gpu-cash-hashrate":   v.GPUCashHashrate(),
+		"gpu-redshift":        v.GPURedshift(),
+	}
+}
+
+func benchmarksToMap(b *sonm.Benchmarks) map[string]uint64 {
+	v := Benchmarks(*b)
+	return v.toMap()
+}
